code-gen/requirements.txt
pip
stringcase
pyyaml
mako
pytesseract
pillow
cloudinary
requests
GitPython
python-dotenv
Jinja2